+++ src/features/library/screens/Library.tsx
@@ -316,7 +316,7 @@ export function Library() {
             </OffsetComponent>
             {triggerGlobalSearchButton}
             {tabs.map((tab) => (
-                <TabPanel key={tab.order} index={tab.order} currentIndex={activeTab.order}>
+                <TabPanel key={tab.id} index={tab.order} currentIndex={activeTab.order}>
                     {tab === activeTab && (
                         <LibraryMangaGrid
                             // the key needs to include filters and query to force a re-render of the virtuoso grid to prevent https://github.com/petyosi/react-virtuoso/issues/1242