+++ src/lib/graphql/download/DownloadFragments.ts
@@ -7,12 +7,14 @@
  */
 
 import gql from 'graphql-tag';
-import { CHAPTER_BASE_FIELDS } from '@/lib/graphql/chapter/ChapterFragments.ts';
+import { SOURCE_BASE_FIELDS } from '@/lib/graphql/source/SourceFragments.ts';
 import { MANGA_BASE_FIELDS } from '@/lib/graphql/manga/MangaFragments.ts';
+import { CHAPTER_BASE_FIELDS } from '@/lib/graphql/chapter/ChapterFragments.ts';
 
 export const DOWNLOAD_TYPE_FIELDS = gql`
-    ${CHAPTER_BASE_FIELDS}
     ${MANGA_BASE_FIELDS}
+    ${SOURCE_BASE_FIELDS}
+    ${CHAPTER_BASE_FIELDS}
 
     fragment DOWNLOAD_TYPE_FIELDS on DownloadType {
         chapter {
@@ -23,6 +25,9 @@ export const DOWNLOAD_TYPE_FIELDS = gql`
         manga {
             ...MANGA_BASE_FIELDS
             downloadCount
+            source {
+                ...SOURCE_BASE_FIELDS
+            }
         }
 
         progress